requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.2
billiard==3.3.0.23
boto==2.42.0
celery==3.1.25
dj-static==0.0.6
Django==1.10.5
django-bootstrap3==7.0.1
django-celery==3.2.1
django-cors-headers==1.3.0
django-filter==0.15.2
django-fontawesome==0.2.6
django-storages==1.5.0
django-ses-gateway==0.1.1
djangorestframework==3.3.3
gunicorn==18.0
kombu==3.0.37
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
pyparsing==2.1.10
pytz==2016.10
PyYAML==3.12
six==1.10.0
static3==0.5.1
vine==1.1.3
boto3==1.0.0
